+++ R/AutoGlmnet.R
@@ -30,7 +30,9 @@ AutoGlmnet = R6Class(
         properties = character(),
         task_types = c("classif", "regr"),
         packages = c("mlr3", "mlr3learners", "glmnet"),
-        devices = "cpu"
+        devices = "cpu",
+        n_cpu = 1L,
+        n_gpu = 0L
       )
     },